+++ reservation-service/src/main/java/kr/or/connect/reservation/service/UserService.java
@@ -3,6 +3,6 @@ package kr.or.connect.reservation.service;
 import kr.or.connect.reservation.domain.User;
 
 public interface UserService {
-	public long addUser(User user);
+	public Integer addUser(User user);
 	public User getUser(String id);
 }
